[Sb refactor] Convert CastTile to use WifiRepository
CastTile was only using a single bit of information from the old data
pipeline. If the QS wifi icon was visible then it would consider "wifi
enabled". This CL sets up a new TileJavaAdapter which can be used to
collect on flows from QS, and wires up the wifi repository for CastTile.

Feature is gated behind the SIGNAL_CALLBACK_DEPRECATION flag. All tests
of the old behavior have been replicated with the flag set to `true`

Test: CastTileTest
Test: manual - check hotspot tile with wifi on/off while connected and
disconnected to hotspot. Verified the same behavior with the flag on and
off
Bug: 291321279
Change-Id: I67755215ed0222c79eaae7dc8ebaf41b4d65e8ad

Change-Id: I54b7af0c9f9c084305e7447051f37e6fd855bfb5