Differential motion fling for platform widgets
This change adds a DifferentialMotionFlingHelper class to control
flinging from differential motion, and uses it in AbsListView and
ScrollView.

The helper class is currently hidden. Unless we have a strong incentive
to open up the API for developers, we will keep the API hidden.

Demo video: http://shortn/_0quHhjPtCV

Interaction with stretch effect is also demonstrated in the demo.

Bug: 293332089
Test: manual, DifferentialMotionFlingHelperTest
Test: no CTS regression for ScrollViewTest, ListViewTest, GridViewTest
Change-Id: I7acd1aa63da4c64d72cca3c8ac6e10dea1f5bd0b