Save BT hearing device event history for HaTS surveys
We'll run surveys on hearing adis and general hearing devices users to
know the satisfaction difference between different kinds of devices.
This can help us prioritize to improve the parts which are less satisfying.

getProfiles() are not availabe right after the bonded event. We need to wait until the device is connected to see if it's a hearing device or not. Delay
the saving time of hearing devices bonded events until the profiles are connected.

Bug: 294627726
Test: atest HearingAidStatsLogUtilsTest
Test: atest CachedBluetoothDeviceTest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:ea207bdb0a87dccd7026cc133b407f3081c7c823)

Change-Id: I70dc1680bea2f1002325c2ba1497dfccb5cf2b48