Fix will log triple tap on shortcut temporary glance mode
The long press after the shortcut case and the triple tap and hold case will both call the same function. This caused the shortcut case to be logged as a triple tap event. I added a check to the condition to ensure that the shortcut case was not logged. This fixed the issue and also fixed the incorrect log enable state.

Bug: 296036296
Test: manual
Test: atest FullScreenMagnificationGestureHandlerTest
Change-Id: I10df9d434c31d7f6d043e9f860eb9bbef3af7615