Adds intent extra to package data clear broadcast during restore operation.
During full restore of a package, Backup service asks AM service to clear app data for the package (to ensure clean state for restore). AM broadcasts ACTION_PACKAGE_DATA_CLEARED intent, which Blockstore receives and it clears associated Blockstore data. This results in Blockstore data not being restored.

Adding an intent extra to the broadcast enabled Blockstore to know that it happened because of a restore, and won't clear it's data. See go/br-blockstore-framework-fix for details.

Note: This change adds intent extra as @hide. b/298012653 tracks the work to add it as a system api for V.

Bug: 243464090
Test: Verified through BroadcastReceiver
Change-Id: I3e90279a3dd2e6753b76527cbdb3f70ecce4db0f