Make volume UI work when using eARC.
Makes the ARC and eARC audio output devices always have the same volume
behavior on TV panels. That is, when AVB is enabled on a TV, both the
ARC and eARC audio output devices adopt absolute volume behavior;
when it is disabled, both adopt full volume behavior.

This allows the existing logic to extend to eARC without significant
changes to the code flow.

Bug: 265683383
Test: atest TvToAudioSystemEarcAvbTest; manual on Gambit
Change-Id: Ida4ba65880e12fd7071579d3e70c611c235fd3ad