Fix DPC test
This test was failing on devices where the configuration for
config_allowAutoBrightnessWhileDozing defaulted or was set to true.

The config should be true, when allowing full automatic screen
brightness when in doze, and therefore, screenoffbrightnesssensor should
be enabled, only when this value is false, since it is used as a
fallback sensor.

The timings of the setting the return value of this testable resource
needs to come before the variable is actually set when creating the DPC,
which is why the DPC needed to be re-initialised in this test.

Bug: 275079012
Test: atest DisplayServiceTests
Change-Id: I2eee26460d11dae82994b648d30e834547e8483a