Include TrustState in TrustManagerService dumpUser
TrustManagerService dumps the raw aggregated trustable value for each
user, which is not useful because it is not the authoritative trust
state for the user, and it's a computed value (i.e., not real state)
that is computed from other state already dumped.  Therefore, dump the
TrustState instead which is real state that is much more useful.

Test: adb shell dumpsys trust
Flag: exempt, straightforward fix that only affects bugreports
Change-Id: I17103be4d0cc2545e3bdcb1656b35f2d287b01d5