Use default ScreenshotController for PROVIDED_IMAGE type
If the screenshot type is PROVIDED_IMAGE, then there is no need to make multiple display screenshots.
In this case, the screenshot controller for the default display is used to save the image and show the UI.

In the future, we might want to show the UI only in the display that actually provided the image, but for now this is not possible, as only the default display can provide the image.

Bug: 290910794
Bug: 297354762
Test: TakeScreenshotExecutorTest
Change-Id: I1e30be46cf405a0e52047946a1e993266bac94dc