Only recompute new config if rotation is changed from settings
It was true because the settings were mixed with the other settings
in PhoneWindowManager that may contain keyboard state change. Since
the settings observer of rotation is extracted to only listen rotation
related changes, it no longer needs to always compute new display
configuration that is either a wasteful invocation or may disturb the
display info by some intermediate states, such as launching app in
different orientation.

Bug: 298331183
Test: AppConfigurationTests#testFixedOrientationWhenRotating
Change-Id: I80abecf55074f10999387ad1c2c7f14ff11c403c