Also catch InvalidDisplayException in SoftInputWindow#show()
This is a follow up to my previous commit [1], which introduced a
state machine to handle inevitable runtime failures in a graceful
manner. Since then, we have started relying more on multi-display,
which sometimes ends up throwing

  WindowManager.InvalidDisplayException

instead of

  WindowManager.BadTokenException.

From IMEs' viewpoints, that they are equally unavoidable. Let's catch
both of them.

 [1]: I1f51ed29df6b22b003ca285698e06b8f3ee5106a
      16891dd1b26dc4b28fee4b9191cc4c562f788e9d

Fix: 295880764
Test: no end-to-end test as we have not figured out how to reproduce
      this issue locally.
Change-Id: I66ba7b465e901e9ec2b31dd0245420513381872d