Catch an exception when loading a bubble icon
If an app gives us a shortcut whose icon can't be looked up, an
exception is thrown. This CL modifies the code to catch an
exception and default to the app icon instead of the shortcut
icon.

Test: manual - modify icon loading code to throw an exception
             - create a bubble
	     => observe there is no crash and the bubble uses
	        the app icon instead
Bug: 300073572

Change-Id: Iedc9825c42ec2f50bb9939e34acd9f1960dd229e
(cherry picked from commit 21973d05ef6f3055636b95cae7e3f24ed09457ba)